PostgreSQL backend
This PR is @cpcloud's work plus fixes for the data load. There is a
test failure on PostgreSQL 9.3 we'll need to investigate and fix.
Closes #720

Author: Phillip Cloud <cpcloud@gmail.com>
Author: Wes McKinney <wesm@apache.org>

Closes #818 from wesm/pg and squashes the following commits:

3528780 [Wes McKinney] Fix up data loading script for postgres, make variance statement work on postgres 9.3 (unverified)
d80f0d0 [Phillip Cloud] Add support for cumulative window functions
0964902 [Phillip Cloud] Make timestamp tests 100% reliable
3a2e618 [Phillip Cloud] Clean up some docs
8a6a4cc [Phillip Cloud] Ensure that we can find NULLs in sets
ae579ce [Phillip Cloud] Make sure that NullLiteral has a value so that we can create a list containing them in SQLAlchemy
ad864b8 [Phillip Cloud] Implement regex extract for pg
a9c3f49 [Phillip Cloud] Add test for regex extract
1eeca39 [Phillip Cloud] No need for parens
368430a [Phillip Cloud] Add postgres to test data generation script
4853357 [Phillip Cloud] Small cleanup [ci skip]
8716d75 [Phillip Cloud] Comment [ci skip]
f1086c3 [Phillip Cloud] DRY it up
7bfdf3f [Phillip Cloud] Implement basic window functions for pg
213e187 [Phillip Cloud] Add rows specification to SQLAlchemy OVER clause
ebdbff4 [Phillip Cloud] Add more window tests
185a76e [Phillip Cloud] Test simple window
3dbb4a1 [Phillip Cloud] Begin window work
43395b2 [Phillip Cloud] XFail so that we know what should work and what should be skipped
5388794 [Phillip Cloud] Fix variance and log
eb2541d [Phillip Cloud] Add tests for math et al
32b17e6 [Phillip Cloud] Use CTEs for better looking SQL
86f45d1 [Phillip Cloud] Fail on null scalars for now
f621530 [Phillip Cloud] Implement find_in_set et al
77473db [Phillip Cloud] Add tests for moar string functions!
1cabc28 [Phillip Cloud] Implement reverse
d1c1db1 [Phillip Cloud] Add test for reverse
7c2e2c0 [Phillip Cloud] Implement string padding
2390eec [Phillip Cloud] Add basic string padding functions
55879f9 [Phillip Cloud] Cleanup and comment
c54d2f8 [Phillip Cloud] Implement coalesce
ef6576d [Phillip Cloud] Add tests for coalesce
25b72d6 [Phillip Cloud] Fail on not testing coalesce
4c40a63 [Phillip Cloud] Use replace instead of str_replace
a5d303c [Phillip Cloud] Use re.escape
f33a35c [Phillip Cloud] WIP: Add postgresql backend

+++ ibis/sql/tests/test_sqlalchemy.py
@@ -100,9 +100,9 @@ class TestSQLAlchemySelect(unittest.TestCase, ExprTestCases):
             # name, type, nullable
             ('smallint', sat.SmallInteger, False, dt.int16),
             ('int', sat.Integer, True, dt.int32),
-            ('integer', sat.INTEGER(), True, dt.int64),
+            ('integer', sat.INTEGER(), True, dt.int32),
             ('bigint', sat.BigInteger, False, dt.int64),
-            ('real', sat.REAL, True, dt.double),
+            ('real', sat.REAL, True, dt.float),
             ('bool', sat.Boolean, True, dt.boolean),
             ('timestamp', sat.DateTime, True, dt.timestamp),
         ]
